+++ examples/org-unit-tree/initially-expanded.js
@@ -1,0 +1,22 @@
+import React from 'react';
+
+import OrgUnitTree from '../../src/org-unit-tree';
+
+
+function InitiallyExpandedExample(props) {
+    const expanded = [
+        'ImspTQPwCqd', // Sierra Leone - root
+        'Vth0fbpFcsO', // Kono - under Sierra Leone
+        'EjnIQNVAXGp', // Mafindor - under Kono
+        'ZiOVcrSjSYe', // Dibia - under Port Loco, which is NOT initially expanded!
+    ];
+
+    return (
+        <div>
+            <OrgUnitTree root={props.root} selected={expanded} initiallyExpanded={expanded} />
+        </div>
+    );
+}
+InitiallyExpandedExample.propTypes = { root: React.PropTypes.any.isRequired };
+
+export default InitiallyExpandedExample;