+++ examples/org-unit-tree/single-selection.js
@@ -1,0 +1,45 @@
+import React from 'react';
+
+import OrgUnitTree from '../../src/org-unit-tree';
+
+class SingleSelectionExample extends React.Component {
+    constructor(props) {
+        super(props);
+
+        this.state = {
+            selected: [],
+        };
+
+        this.handleClick = this.handleClick.bind(this);
+    }
+
+    handleClick(event, orgUnit) {
+        this.setState(state => {
+            if (state.selected[0] === orgUnit.id) {
+                return { selected: [] };
+            }
+
+            return { selected: [orgUnit.id] };
+        });
+    }
+
+    render() {
+        const selStyle = {
+            borderTop: '1px solid #eeeeee',
+            margin: '16px -16px 0',
+            padding: '16px 16px 0',
+        };
+        return (
+            <div>
+                <OrgUnitTree
+                    root={this.props.root}
+                    onClick={this.handleClick}
+                    selected={this.state.selected}
+                />
+                <div style={selStyle}>Selected: {this.state.selected}</div>
+            </div>
+        );
+    }
+}
+
+export default SingleSelectionExample;