+++ src/data-table/DataTable.component.js
@@ -35,19 +35,7 @@ const DataTable = React.createClass({
         };
     },
 
-    render() {
-        const headers = this.state.columns.map((headerName, index) => {
-            return (<DataTableHeader key={index} isOdd={Boolean(index % 2)} name={headerName} />);
-        });
-
-        const dataRows = [];
-        let dataRowsId = 0;
-        let dataRowsSource;
-        for (dataRowsSource of this.state.dataRows) {
-            dataRowsId++;
-            dataRows.push(<DataTableRow key={dataRowsId} dataSource={dataRowsSource} columns={this.state.columns} isActive={this.state.activeRow === dataRowsId} itemClicked={this.handleRowClick} primaryClick={this.props.primaryAction || (() => {})} />);
-        }
-
+    renderContextMenu() {
         const actionAccessChecker = (this.props.isContextActionAllowed && this.props.isContextActionAllowed.bind(null, this.state.activeRow)) || (() => true);
 
         const actionsToShow = Object.keys(this.props.contextMenuActions || {})
@@ -58,31 +46,66 @@ const DataTable = React.createClass({
             }, {});
 
         return (
-           <div className="data-table" onClick={this.hideContextMenu} onMouseLeave={this.hideContextMenu}>
+                <DataTableContextMenu
+                    target={this.state.contextMenuTarget}
+                    onRequestClose={this._hideContextMenu}
+                    actions={actionsToShow}
+                    activeItem={this.state.activeRow}
+                />
+        );
+    },
+
+    renderHeaders() {
+        return this.state.columns.map((headerName, index) => {
+            return (
+                <DataTableHeader key={index} isOdd={Boolean(index % 2)} name={headerName} />
+            );
+        });
+    },
+
+    renderRows() {
+        return this.state.dataRows
+            .map((dataRowsSource, dataRowsId) => {
+                return (
+                    <DataTableRow
+                        key={dataRowsId}
+                        dataSource={dataRowsSource}
+                        columns={this.state.columns}
+                        isActive={this.state.activeRow === dataRowsId}
+                        itemClicked={this.handleRowClick}
+                        primaryClick={this.props.primaryAction || (() => {})}
+                    />
+                );
+            });
+    },
+
+    render() {
+        return (
+           <div className="data-table">
                <div className="data-table__headers">
-                    {headers}
+                    {this.renderHeaders()}
+                    <DataTableHeader />
                </div>
                <div className="data-table__rows">
-                   {dataRows}
+                   {this.renderRows()}
                </div>
-               {this.state.activeRow && this.props.contextMenuActions ? <DataTableContextMenu actions={actionsToShow || {}} activeItem={this.state.activeRow} coords={this.state.contextMenuCoords} icons={this.props.contextMenuIcons} /> : undefined}
+               {this.renderContextMenu()}
            </div>
         );
     },
 
     handleRowClick(event, rowSource) {
         this.setState({
-            contextMenuCoords: {
-                Y: event.clientY + window.scrollY - 25,
-                X: event.clientX - 25,
-            },
+            contextMenuTarget: event.currentTarget,
+            showContextMenu: true,
             activeRow: rowSource !== this.state.activeRow ? rowSource : undefined,
         });
     },
 
-    hideContextMenu() {
+    _hideContextMenu() {
         this.setState({
             activeRow: undefined,
+            showContextMenu: false,
         });
     },
 });