+++ test/group-editor/GroupEditor.component.test.js
@@ -1,0 +1,5 @@
+import GroupEditor from '../../src/group-editor/GroupEditor.component';
+
+xdescribe('GroupEditor component', () => {
+    it('has no tests');
+});